Add files via upload
This is the basic concept of the website we were building as a group. The main focus was building a website for a hire company specifically for a cab, which meant we had to have a register system in place as well as a payment system and booking system which are vital.